+++ np.Tidy/CHANGELOG.md
@@ -1,6 +1,12 @@
 # 🧹 Tidy Up Changelog
 See Plugin [README](https://github.com/NotePlan/plugins/blob/main/np.Tidy/README.md) for full details on the available commands and use from callbacks and templates.
 
+## [0.6.0] - 2023-06-???
+- ??? new **/List conflicted notes** command that creates a new NP note that lists all your notes on your current device with file-level conflicts, along with summary details about them
+- new **/Remove blank notes** command will delete any completely blank notes, or just with a starting '#' character
+- improve display of duplicate notes that are empty
+- improved/fixed display of progress dialogs
+
 ## [0.5.0] - 2023-06-12
 - new **/List duplicate notes** command that creates a new NP note that lists all your notes with identical titles, along with summary details about those potential duplicates